Can find a path from one point on the map to another, given a starting point and a goal point.
Subscribes to goal and initial pose topics, and uses data received to plan a path from the specified start point to the goal point. Method for adding a single node to the graph in a structured way added to the graph object.
Modified methods for adding vertices to the graph. Will stop attempting to add if a node has no connections or if the node has already been connected to the node that the connection is being made to. This is necessary since the method will now add the connection information on both sides of the edge. If node 1 connects to node 2, node 2 will be connected to node 1 also, whereas before node 2 had to redo checks for everything before having node 1 added to its connections.